# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [stock]
      summary: Get a vehicle
      parameters:
        - in: path
          name: id
          required: true
          description: Database identifier of the vehicle.
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Vehicle not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /reservations:
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Reserve a vehicle online
      description: Creates a real reservation and lead for an available vehicle. Payment is simulated; no money is received. Repeating the same idempotency key returns the original reservation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnlineReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OnlineReservation" }
        "200":
          description: Previously created reservation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OnlineReservation" }
        "400":
          description: Invalid reservation details
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: Vehicle not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: Vehicle unavailable, reservations disabled, or price or terms changed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "500":
          description: Unable to reserve vehicle
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List online reservations for staff
      description: Requires an authenticated staff member and returns only the configured dealership's reservations.
      responses:
        "200":
          description: Online reservations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OnlineReservations" }
        "401":
          description: Staff sign-in required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: Staff access required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "500":
          description: Unable to load reservations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /reservations/{id}/cancel:
    post:
      operationId: cancelReservation
      tags: [reservations]
      summary: Cancel an online reservation
      description: Requires staff access. Releases the vehicle only when this reservation owns its current hold.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$' }
      responses:
        "200":
          description: Cancelled reservation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OnlineReservation" }
        "401":
          description: Staff sign-in required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "403":
          description: Staff access required
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: Reservation not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: Reservation cannot be cancelled in its current state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "500":
          description: Unable to cancel reservation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
...
components:
  schemas:
    OnlineReservationPartExchange:
      type: object
      additionalProperties: false
      properties:
        registration: { type: string, minLength: 2, maxLength: 16 }
        mileage: { type: number, multipleOf: 1, minimum: 0, maximum: 1000000 }
      required: [registration, mileage]
    OnlineReservationInput:
      type: object
      additionalProperties: false
      properties:
        vehicleId: { type: string, pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$' }
        idempotencyKey: { type: string, pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$' }
        customerName: { type: string, minLength: 2, maxLength: 120 }
        email: { type: string, maxLength: 254, pattern: '^[^\s@]+@[^\s@]+\.[^\s@]+$' }
        phone: { type: string, minLength: 5, maxLength: 40 }
        expectedPricePence: { type: number, multipleOf: 1, minimum: 1, maximum: 2147483647 }
        expectedDepositPence: { type: number, multipleOf: 1, minimum: 100, maximum: 1000000 }
        termsAccepted: { type: boolean, enum: [true] }
        terms: { type: string, minLength: 1, maxLength: 4000 }
        partExchange: { $ref: "#/components/schemas/OnlineReservationPartExchange" }
      required: [vehicleId, idempotencyKey, customerName, email, phone, expectedPricePence, expectedDepositPence, termsAccepted, terms]
    OnlineReservation:
      type: object
      additionalProperties: false
      properties: &onlineReservationProperties
        id: { type: string, pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$' }
        reference: { type: string }
        vehicleId: { type: string }
        vehicleTitle: { type: string }
        vehicleRegistration: { type: [string, "null"] }
        depositPence: { type: number, multipleOf: 1, minimum: 100, maximum: 1000000 }
        amountReceivedPence: { type: number, multipleOf: 1, enum: [0] }
        paymentStatus: { type: string, enum: [simulated] }
        status: { type: string, enum: [reserved, cancelled] }
        createdAt: { type: string, format: date-time }
      required: [id, reference, vehicleId, vehicleTitle, depositPence, amountReceivedPence, paymentStatus, status, createdAt]
    StaffOnlineReservation:
      type: object
      additionalProperties: false
      properties:
        <<: *onlineReservationProperties
        customerName: { type: string }
        email: { type: string }
        phone: { type: string }
        leadId: { type: [string, "null"] }
      required: [id, reference, vehicleId, vehicleTitle, depositPence, amountReceivedPence, paymentStatus, status, createdAt, customerName, email, phone, leadId]
    OnlineReservations:
      type: object
      additionalProperties: false
      properties:
        reservations:
          type: array
          items: { $ref: "#/components/schemas/StaffOnlineReservation" }
      required: [reservations]
...
    Vehicle:
      type: object
      additionalProperties: false
      properties:
        <<: *importedVehicleProperties
        id:
          type: string
          minLength: 1
        inventoryStatus:
          type: string
          readOnly: true
          enum: [available, reserved, sold, hidden, archived]
      required:
        - id
        - advertId
        - title
        - variant
        - make
        - model
        - trim
        - year
        - price
        - priceType
        - currency
        - mileage
        - mileageText
        - registration
        - registrationBand
        - plate
        - vrm
        - vrmVerified
        - fuel
        - transmission
        - bodyType
        - engineSize
        - engineCC
        - doors
        - seats
        - colour
        - emissionClass
        - drivetrain
        - owners
        - writeOffCategory
        - advertUrl
        - dealerName
        - dealerLocation
        - imageCount
        - heroImage
        - images
        - specifications
        - sourceExtras
...
    ApiError:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
          minLength: 1
      required:
        - error
...